backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
google-genai>=0.3.0
requests>=2.32.0
elevenlabs==0.2.27
playwright==1.41.0

